Moved llama logic to llama_pipeline
Did a very basic version of OBP and llama.
You can choose llama as hyperprompt option.
It will randomly pick a llama.txt file and add it when chosen.
